fix(router): rename routes
